+++ app/test/search/mem_index_test.dart
@@ -103,12 +103,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {
-            'package': 'async',
-            'score': closeTo(0.71, 0.01),
-          },
-        ],
         'highlightedHit': {'package': 'async'},
         'sdkLibraryHits': [],
         'packageHits': [],
@@ -121,12 +115,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {
-            'package': 'http',
-            'score': closeTo(0.85, 0.01),
-          },
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {
@@ -143,12 +131,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {
-            'package': 'chrome_net',
-            'score': closeTo(0.29, 0.01),
-          },
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {
@@ -165,10 +147,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 2,
-        'packages': [
-          {'package': 'http', 'score': closeTo(0.70, 0.01)},
-          {'package': 'async', 'score': closeTo(0.62, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http', 'score': closeTo(0.70, 0.01)},
@@ -183,9 +161,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'async', 'score': closeTo(0.29, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'async', 'score': closeTo(0.29, 0.01)},
@@ -199,12 +174,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {
-            'package': 'chrome_net',
-            'score': closeTo(0.54, 0.1),
-          },
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {
@@ -221,10 +190,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 2,
-        'packages': [
-          {'package': 'http', 'score': 0.5},
-          {'package': 'chrome_net', 'score': closeTo(0.11, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http', 'score': 0.5},
@@ -239,11 +204,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 3,
-        'packages': [
-          {'package': 'async'},
-          {'package': 'http'},
-          {'package': 'chrome_net'},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'async'},
@@ -259,11 +219,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 3,
-        'packages': [
-          {'package': 'http'},
-          {'package': 'async'},
-          {'package': 'chrome_net'},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http'},
@@ -279,10 +234,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 2,
-        'packages': [
-          {'package': 'http'},
-          {'package': 'chrome_net'},
-        ],
         'highlightedHit': {'package': 'http'},
         'sdkLibraryHits': [],
         'packageHits': [
@@ -302,10 +253,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 2,
-        'packages': [
-          {'package': 'http'},
-          {'package': 'async'},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http'},
@@ -320,11 +267,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 3,
-        'packages': [
-          {'package': 'async', 'score': 0.8},
-          {'package': 'http', 'score': 0.7},
-          {'package': 'chrome_net', 'score': 0.0},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'async', 'score': 0.8},
@@ -340,11 +282,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 3,
-        'packages': [
-          {'package': 'http', 'score': 10.0},
-          {'package': 'async', 'score': 1.0},
-          {'package': 'chrome_net', 'score': 0.0},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http', 'score': 10.0},
@@ -360,11 +297,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 3,
-        'packages': [
-          {'package': 'http', 'score': 110.0},
-          {'package': 'async', 'score': 10.0},
-          {'package': 'chrome_net', 'score': 0.0},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http', 'score': 110.0},
@@ -380,10 +312,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 2,
-        'packages': [
-          {'package': 'http', 'score': closeTo(0.96, 0.01)},
-          {'package': 'async', 'score': closeTo(0.70, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http', 'score': closeTo(0.96, 0.01)},
@@ -398,9 +326,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'chrome_net', 'score': closeTo(0.54, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'chrome_net', 'score': closeTo(0.54, 0.01)},
@@ -414,10 +339,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 2,
-        'packages': [
-          {'package': 'http', 'score': closeTo(0.96, 0.01)},
-          {'package': 'chrome_net', 'score': closeTo(0.54, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http', 'score': closeTo(0.96, 0.01)},
@@ -432,9 +353,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'http', 'score': closeTo(0.85, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http', 'score': closeTo(0.85, 0.01)},
@@ -448,9 +366,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'http', 'score': closeTo(0.96, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http', 'score': closeTo(0.96, 0.01)},
@@ -464,7 +379,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 0,
-        'packages': [],
         'sdkLibraryHits': [],
         'packageHits': [],
       });
@@ -476,9 +390,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'async', 'score': closeTo(0.70, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'async', 'score': closeTo(0.70, 0.01)},
@@ -492,7 +403,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 0,
-        'packages': [],
         'sdkLibraryHits': [],
         'packageHits': [],
       });
@@ -504,9 +414,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'async', 'score': closeTo(0.70, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'async', 'score': closeTo(0.70, 0.01)},
@@ -523,10 +430,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 2,
-        'packages': [
-          {'package': 'http', 'score': closeTo(0.96, 0.01)},
-          {'package': 'async', 'score': closeTo(0.70, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http', 'score': closeTo(0.96, 0.01)},
@@ -541,7 +444,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 0,
-        'packages': [],
         'sdkLibraryHits': [],
         'packageHits': [],
       });
@@ -553,10 +455,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(result)), {
         'timestamp': isNotNull,
         'totalCount': 2,
-        'packages': [
-          {'package': 'http', 'score': closeTo(0.96, 0.01)},
-          {'package': 'async', 'score': closeTo(0.70, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http', 'score': closeTo(0.96, 0.01)},
@@ -572,9 +470,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(composable)), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'http', 'score': closeTo(0.88, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http', 'score': closeTo(0.88, 0.01)},
@@ -586,11 +481,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(library)), {
         'timestamp': isNotNull,
         'totalCount': 3,
-        'packages': [
-          {'package': 'chrome_net', 'score': closeTo(0.88, 0.01)},
-          {'package': 'async', 'score': closeTo(0.88, 0.01)},
-          {'package': 'http', 'score': closeTo(0.72, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'chrome_net', 'score': closeTo(0.88, 0.01)},
@@ -607,9 +497,6 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       expect(json.decode(json.encode(both)), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'http', 'score': closeTo(0.54, 0.01)},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'http', 'score': closeTo(0.54, 0.01)},
@@ -619,13 +506,13 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
 
     test('non-word query with default order', () async {
       final rs = await index.search(ServiceSearchQuery.parse(query: '='));
-      expect(rs.packages, isEmpty);
+      expect(rs.isEmpty, isTrue);
     });
 
     test('non-word query with text order', () async {
       final rs = await index.search(
           ServiceSearchQuery.parse(query: '=', order: SearchOrder.text));
-      expect(rs.packages, isEmpty);
+      expect(rs.isEmpty, isTrue);
     });
   });
 
@@ -640,15 +527,15 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       ));
       final match = await index.search(
           ServiceSearchQuery.parse(query: 'app', order: SearchOrder.text));
-      expect(match.packages.map((e) => e.toJson()), [
-        {'package': 'app', 'score': 1.0},
+      expect(match.allPackageHits.map((e) => e.toJson()), [
+        {'package': 'app'},
         {'package': 'apps', 'score': 1.0},
       ]);
       final match2 = await index.search(
           ServiceSearchQuery.parse(query: 'apps', order: SearchOrder.text));
-      expect(match2.packages.map((e) => e.toJson()), [
+      expect(match2.allPackageHits.map((e) => e.toJson()), [
+        {'package': 'apps'},
         {'package': 'app', 'score': 1.0},
-        {'package': 'apps', 'score': 1.0},
       ]);
     });
 
@@ -662,14 +549,14 @@ server.dart adds a small, prescriptive server (PicoServer) that can be configure
       ));
       final match = await index.search(
           ServiceSearchQuery.parse(query: 'app', order: SearchOrder.text));
-      expect(match.packages.map((e) => e.toJson()), [
-        {'package': 'app', 'score': 1.0},
+      expect(match.allPackageHits.map((e) => e.toJson()), [
+        {'package': 'app'},
         {'package': 'appz', 'score': 0.75},
       ]);
       final match2 = await index.search(
           ServiceSearchQuery.parse(query: 'appz', order: SearchOrder.text));
-      expect(match2.packages.map((e) => e.toJson()), [
-        {'package': 'appz', 'score': 1.0},
+      expect(match2.allPackageHits.map((e) => e.toJson()), [
+        {'package': 'appz'},
         // would be nice if `app` would show up here
       ]);
     });