+++ app/test/search/scope_specificity_test.dart
@@ -67,24 +67,6 @@ void main() {
       expect(json.decode(json.encode(withoutPlatform)), {
         'timestamp': isNotNull,
         'totalCount': 4,
-        'packages': [
-          {
-            'package': 'json_0',
-            'score': closeTo(0.4, 0.01),
-          },
-          {
-            'package': 'json_1',
-            'score': closeTo(0.4, 0.01),
-          },
-          {
-            'package': 'json_2',
-            'score': closeTo(0.4, 0.01),
-          },
-          {
-            'package': 'json_3',
-            'score': closeTo(0.4, 0.01),
-          },
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {
@@ -117,16 +99,6 @@ void main() {
       expect(json.decode(json.encode(withPlatform)), {
         'timestamp': isNotNull,
         'totalCount': 2,
-        'packages': [
-          {
-            'package': 'json_1',
-            'score': closeTo(0.40, 0.01),
-          },
-          {
-            'package': 'json_3',
-            'score': closeTo(0.36, 0.01),
-          },
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {