+++ app/test/search/tags_test.dart
@@ -22,7 +22,6 @@ void main() {
       expect(rs.toJson(), {
         'timestamp': isNotNull,
         'totalCount': 0,
-        'packages': [],
         'sdkLibraryHits': [],
         'packageHits': [],
       });
@@ -39,9 +38,6 @@ void main() {
       expect(json.decode(json.encode(rs.toJson())), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'pkg1', 'score': isNotNull},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'pkg1', 'score': isNotNull},
@@ -61,9 +57,6 @@ void main() {
       expect(json.decode(json.encode(rs.toJson())), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'pkg2', 'score': isNotNull},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'pkg2', 'score': isNotNull},
@@ -83,9 +76,6 @@ void main() {
       expect(json.decode(json.encode(rs.toJson())), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'pkg1', 'score': isNotNull},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'pkg1', 'score': isNotNull},
@@ -107,10 +97,6 @@ void main() {
       expect(json.decode(json.encode(rs.toJson())), {
         'timestamp': isNotNull,
         'totalCount': 2,
-        'packages': [
-          {'package': 'pkg1', 'score': isNotNull},
-          {'package': 'pkg2', 'score': isNotNull},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'pkg1', 'score': isNotNull},
@@ -133,9 +119,6 @@ void main() {
       expect(json.decode(json.encode(rs.toJson())), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'pkg2', 'score': isNotNull},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'pkg2', 'score': isNotNull},
@@ -157,9 +140,6 @@ void main() {
       expect(json.decode(json.encode(rs.toJson())), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'pkg1', 'score': isNotNull},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'pkg1', 'score': isNotNull},
@@ -179,9 +159,6 @@ void main() {
       expect(json.decode(json.encode(rs.toJson())), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'pkg2', 'score': isNotNull},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'pkg2', 'score': isNotNull},
@@ -202,9 +179,6 @@ void main() {
       expect(json.decode(json.encode(rs.toJson())), {
         'timestamp': isNotNull,
         'totalCount': 1,
-        'packages': [
-          {'package': 'pkg1', 'score': isNotNull},
-        ],
         'sdkLibraryHits': [],
         'packageHits': [
           {'package': 'pkg1', 'score': isNotNull},